+++ src/fr/helmdefense/model/entities/attackers/bosses/Boss.java
@@ -1,7 +1,7 @@
 package fr.helmdefense.model.entities.attackers.bosses;
 
 import fr.helmdefense.model.entities.attackers.Attacker;
-import fr.helmdefense.model.entities.utils.Location;
+import fr.helmdefense.model.entities.utils.coords.Location;
 
 public  abstract class Boss extends Attacker {
 	
@@ -9,7 +9,7 @@ public  abstract class Boss extends Attacker {
 		super(location);
 	}
 	
-	public Boss(int x, int y) {
+	public Boss(double x, double y) {
 		this(new Location(x, y));
 	}
