+++ src/fr/helmdefense/model/entities/defenders/Catapult.java
@@ -1,6 +1,6 @@
 package fr.helmdefense.model.entities.defenders;
 
-import fr.helmdefense.model.entities.utils.Location;
+import fr.helmdefense.model.entities.utils.coords.Location;
 
 public class Catapult extends Defender {
 	
@@ -8,7 +8,7 @@ public class Catapult extends Defender {
 		super(loc);
 	}
 	
-	public Catapult(int x, int y) {
+	public Catapult(double x, double y) {
 		this(new Location(x,y));
 	}
