+++ src/fr/helmdefense/model/entities/defenders/heros/Aragorn.java
@@ -1,6 +1,6 @@
 package fr.helmdefense.model.entities.defenders.heros;
 
-import fr.helmdefense.model.entities.utils.Location;
+import fr.helmdefense.model.entities.utils.coords.Location;
 
 public class Aragorn extends Hero {
 	
@@ -8,7 +8,7 @@ public class Aragorn extends Hero {
 		super(loc);
 	}
 	
-	public Aragorn(int x, int y) {
+	public Aragorn(double x, double y) {
 		this(new Location(x, y));
 	}
 }
