+++ src/fr/helmdefense/model/entities/defenders/heros/Hero.java
@@ -1,11 +1,11 @@
 package fr.helmdefense.model.entities.defenders.heros;
 
 import fr.helmdefense.model.entities.defenders.Defender;
-import fr.helmdefense.model.entities.utils.Location;
+import fr.helmdefense.model.entities.utils.coords.Location;
 
 public abstract class Hero extends Defender {
 
-	public Hero(int x, int y) {
+	public Hero(double x, double y) {
 		this(new Location(x, y));
 	}
